Seperating the Discord Integration as it's scope begins to outstrip the original app (now in commands/mtgsearch.js)
