+++ .gitignore
@@ -1,3 +1,4 @@
 /node_modules
 npm-debug.log
-/discord_integration/config.json
+/config.json
+.env
