+++ commands/ping.js
@@ -1,0 +1,11 @@
+const { SlashCommandBuilder } = require(discord.js);
+
+module.exports = {
+    data: new SlashCommandBuilder()
+    .setName('ping')
+    .setDescription('Replies with Pong!'),
+    
+    async execute(interaction) {
+        await interaction.reply('Pong Dude!');
+    },
+};
