+++ package.json
@@ -10,6 +10,7 @@
   "license": "ISC",
   "dependencies": {
     "discord.js": "^14.7.1",
+    "dotenv": "^16.0.3",
     "inquirer": "^6.3.1",
     "mtgsdk": "^1.0.0",
     "nodemon": "^2.0.20"
